5/3
I was tardy for class today by 20 minutes due to an excused doctors appointment. I finished up with the lecture styles and started working on Lab 5. I made the pages, figured out some of the buttons, downloaded and put in most of the pictures, and made the frames and pages. Will finish tonight.