5/4
I was 30 minutes late to class because my parking lot couldn't be used so I had to walk 5+ blocks to get here. Today I managed to figure out how to put pictures into button names, and put the pictures and content into each of the buttons on the main page.