5/11
Today i came up an idea with my app for the final project. The idea was to have a yugioh card app, where the app would remember the users cards by saving them to a text file. It would also display all the cards by type and by rarity. It would also finally use an api to calculate the total price of all the cards. I found the api that might work today