Updatee
Today I worked on my spec for the project. I finished writing the first part of it following the instructions on the word document. I am planning on doing a trading card game collection tracker, that pulls prices from an online price api. By doing this, it can show the user the price of their collection, while also being able to display the user's cards to them.