5/15
Working on the spec for the project. Looking into details on how the project will go. I already explained project in previous commit. Today I tried to see if rerouting the source file of downloaded images is possible. Trying to make the scope of the project smaller so that I can finish.